% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{AreaClass}
\alias{AreaClass}
\title{Defines the area class}
\usage{
AreaClass(area)
}
\arguments{
\item{backlog}{A data frame containing columns titled \code{timestep},
\code{grade} and \code{area}. There can be other columns.}
}
\description{
The area class is a long data format holding area condition summaries
suitable for passing to \code{ggplot2}.
}
