% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areafy.R
\name{areafy2}
\alias{areafy2}
\title{Estimate the unit_area for an observation for a blockbuster tibble with readable code.}
\usage{
areafy2(blockbuster_initial_state, unit_area_methods = "PDS")
}
\arguments{
\item{blockbuster_initial_state}{a blockbuster dataframe or tibble without \code{unit_area}.}

\item{unit_area_methods}{a string to specify whether "PDS" or "CDC". Currently only "PDS" supported.}
}
\value{
a blockbuster tibble with the unit_area estimated for each row.
}
\description{
This should be used to provide the initial estimates of the unit_area
of each element-sub-element construction type in the dataframe found in the
R script 01_read_and_tidy_data.R.
}
\details{
It may be preferred over the deprecated \code{areafy} as its easier to read,
however as \code{\link[dplyr]{case_when}} is experimental, areafy may be preferred.  
It will not need to be called after deterioration and different grades of the elementid appear,
as the unit_area will be calculated by the transition from the superior grade condition.
As most unit area calculation methods are to use the gifa,
 we do this first and then specify the rarer \code{unit_area}
calculations or estimation methods. A window or door is assumed to have an area of one 
square metre. See the data-raw file blockbuster_unit_quantity_method.csv for method details.
The first areafy function used nested ifelse statements and was difficult to read. This version is
an attempt to tidy that and make the code more human readable for debugging.
}
\examples{
pds_data <- areafy2(blockbuster_pds[1, ])$unit_area 
pds_data == blockbuster_pds[1, ]$unit_area  
}

