% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{BacklogE}
\alias{BacklogE}
\title{Compute total repair cost of blocks at grade D using block-level data.}
\usage{
BacklogE(block.data)
}
\arguments{
\item{block.data}{A \code{\link{block}} object.}
}
\value{
A number.
}
\description{
Compute total repair cost of blocks at grade D using block-level data.
}
