% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deteriorate.R
\name{blockbust}
\alias{blockbust}
\title{The det_erioration of more than one blockbuster rows through one time period.}
\usage{
blockbust(blockbuster_tibble)
}
\arguments{
\item{blockbuster_tibble}{a blockbuster dataframe or tibble.}
}
\value{
The input blockbuster_tibble deteriorated through one timestep containing
the \code{unit_area} and condition of the \code{element sub_element constr_type}  
combination after one time period. Duplicating all other variables and values.
The timestep also increases by one. The output tibble can be up to twice the number
of rows of the input tibble. This function is built using \code{purrr::map} to vectorise
the \code{\link{det_eriorate}} function.
}
\description{
The det_erioration of more than one blockbuster rows through one time period.
}
\examples{
\dontrun{
one_year_later <- blockbust(dplyr::filter(blockbuster_pds, buildingid == 127617))
}
}
