% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockbuster.R
\name{Blockbuster}
\alias{Blockbuster}
\title{Blockbuster model of school estate deterioration}
\usage{
Blockbuster(element.data, block.data = NULL, forecast.horizon = 1,
  rebuild.money = 0, repair.money = 0, block.rebuild.cost = 2000,
  inflation = 1, block.det.data = blockbuster_det_data,
  block.repair.costs = blockbuster_pds_repair_costs, save = TRUE,
  filelabel = "blockbuster_output", path = "./output/",
  grade.order = c("D", "C", "B"))
}
\arguments{
\item{element.data}{An \code{\link{element}} object containing the initial
state of the block components.}

\item{block.data}{(optional) A \code{\link{block}} object containing the
initial state of the blocks. If not supplied, this will be created from the
element.data.}

\item{forecast.horizon}{A number. How many timesteps are to be simulated.
Default value is 1.}

\item{rebuild.money}{A vector of numbers indicating the available money to
spend on rebuilding blocks each timestep. Default value is zero.}

\item{repair.money}{A vector of numbers indicating the available money to
spend on repairing components each timestep. Default value is zero}

\item{block.rebuild.cost}{(optional) The unit cost (per m^2^) of rebuilding blocks.
Default value is 2000. This is required if the \code{block.data} argument is not
supplied.}

\item{inflation}{A vector of numbers indicating the inflation rate to apply
to repair and rebuild costs each timestep.  If a vector of length one is
supplied it will be used as the inflation rate for all timesteps. Default
is 1, i.e. no inflation.}

\item{block.det.data}{(optional) An object of the same form as
\code{block_det_data} that is used to provide deterioration rates.}

\item{block.repair.costs}{(optional) an object of the same form as
\code{\link{blockbuster_pds_repair_costs}} that is used to provide repair
costs.}

\item{save}{(optional) Logical. If \code{TRUE} (the default behaviour), then
the block and element states are saved to disc at each timestep.}

\item{filelabel}{(optional) Character. The start of the file names used to save
the interim outputs. Default is "blockbuster_output".}

\item{path}{(optional) Character. The path to the folder which stores the
outputs. If the folder this points to does not exist then the simulation will
fail. Default is "./output/".}

\item{grade.order}{(Optional) A list of the three character strings \code{"D"},
\code{"C"} and \code{"B"} in any order.  This determines the priority for
repairs.  The first character gives the first grade that will be repaired.
By default, D is repaired first, then C, then B.}
}
\value{
A list.  Each list entry contains the \code{\link{block}} and
\code{\link{element}} objects that store the condition of the blocks and
components at each timestep, with the initial conditions stored in the first
list entry.
}
\description{
High level modelling of element and block objects through time that considers
repairs and rebuilding interventions on the condition of the modelled
building components. It is composed of many smaller functions that
deteriorate, rebuild and repair the components.
Outputs a list of block and element objects detailing the condition of the
modelled blocks and components at each timestep.  Information is stored as
\code{\link{block}} and \code{\link{element}} objects. The states at each
timestep and the final output are saved as files specified by the
\code{filelabel} and \code{path} parameters.  The saving of outputs to disc is
mandatory.  It saves memory space for large samples and simulations and the
final output is collated from the individual files rather than being stored
in memory.
}
\details{
At each timestep the components are degraded at the appropriate rates.
Inflation (set by the user using the \code{inflation} parameter) is then
applied to repair and rebuild costs.  After this, blocks are
rebuilt in an order determined by a cost benefit calculation which weighs the
cost of repairing all components in a block with the cost of rebuilding the
block. Once this is done, the model then decides on which components need
repairing.  Grade D components are repaired first, in order of descending
cost.  Then Grade C components are repaired and then, if the budget allows,
Grade B.

The degradation rates of a component are based on transition rates
from EC Harris and expert opinion within DfE.  Alternate transition rates can
be passed to the function using the \code{block.repair.costs} parameter after
manually editing the block_det_data object created by the library.

Repair costs per unit area for each component are pulled from the
\code{link{blockbuster_pds_repair_costs}} data table created by the
blockbuster package.  Repair costs can be manually passed to the blockbuster
function using the \code{block.repair.costs} argument, respecting the format
of the default parameter.

Rebuild costs per unit area are user-input as parameter
\code{block.rebuild.cost} and individiual block rebuild costs generated from
this and the GIFA (Gross Internal Floor Area) of the block.

The \code{rebuild.money} and \code{repair.money} arguments provide the
budgets for rebuilding and repairing during each timestep respectively.
}
\examples{
TODO
}
