% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockbuster_det_data}
\alias{blockbuster_det_data}
\title{Deterioration rates of all blockbuster element-sub-element-construction types.}
\format{A tibble with 147 rows and 9 variables:}
\usage{
blockbuster_det_data
}
\description{
A dataset containing the Deterioration rates of all element by sub-element
by construction type combinations. These rates were provided by the
consultant EC Harris and are necessary for predicting the
deterioration of each element by sub-element
by construction type combination through time as in the
 \code{det_eriorate} function in this package.
}
\keyword{...}
\keyword{\describe{}
\keyword{\item{concated_det}{A}
\keyword{\item{const_type}{Construction}
\keyword{\item{element}{Building}
\keyword{\item{sub_element}{}}
\keyword{\item{transition}
\keyword{a}
\keyword{and}
\keyword{b}
\keyword{b,}
\keyword{c}
\keyword{c,}
\keyword{classification}}
\keyword{const_type.}}
\keyword{d}
\keyword{data}
\keyword{e.}}
\keyword{element}
\keyword{of}
\keyword{pasting}
\keyword{rates}{From}
\keyword{sub-element}
\keyword{taxonomic}
\keyword{the}
\keyword{to}
\keyword{together}
\keyword{type}}
\keyword{}}

