% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_input.R
\name{blockbuster_excel}
\alias{blockbuster_excel}
\title{Run the deterioration model from an excel sheet}
\usage{
blockbuster_excel(path)
}
\arguments{
\item{path}{character. The path where you want to save the output files}
}
\value{
The locations of the output excel workbook and word document, along
with the corresponding files in the temp folder in case something goes wrong
with file permissions.
}
\description{
Run the deterioration model from an excel sheet
}
