% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockbuster_mc_list}
\alias{blockbuster_mc_list}
\title{A markovList of markovchain objects containing the Deterioration rates.}
\format{A S4 markovList of 147 markovchain S4 objects:
\describe{
  \item{markovchains}{A list of 6 by 6 markovchain transition matrices}
  \item{name}{Name of the entire list.}
  ...
}}
\usage{
blockbuster_mc_list
}
\description{
A markovList containing the Deterioration rates of all element by sub-element
by construction type combinations as markovchain objects. 
These rates were provided by the
consultant EC Harris and are necessary for predicting the
deterioration of each element by sub-element
by construction type combination through time as in the
 \code{deteriorate} function in this package. Similar to
 the \code{blockbuster_det_data} but as a S4 rather than S3.
 To access slots of a S4 object use "@" or \code{slot}()
}
\keyword{datasets}

