% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockbuster_sim.R
\name{blockbuster_sim}
\alias{blockbuster_sim}
\title{RAM efficient version of blockbuster designed to scale.}
\usage{
blockbuster_sim(blockbuster_tibble, forecast_horizon, rebuild_monies = 0,
  repair_monies = 0, rebuild_cost_rate = 1274, output_dir = "./output/",
  output_filename = paste0(Sys.Date(), "-blockbuster_sim"))
}
\arguments{
\item{blockbuster_tibble}{a blockbuster dataframe or tibble.}

\item{forecast_horizon}{an integer for the number of timesteps to model deterioration over.}

\item{rebuild_monies}{a numeric vector of length equal to the \code{forecast_horizon} or one.}

\item{repair_monies}{a numeric vector of length equal to the \code{forecast_horizon} or one.}

\item{rebuild_cost_rate}{a numeric vector of length equal to the \code{forecast_horizon} or one.}

\item{output_dir}{a character string of where to write each timestep's tibble to.}

\item{output_filename}{a character string of the desired filename. Default is 
system date and the timestep.}
}
\value{
An object of class \code{blockbuster_list} 
of n plus one tibbles (where n is the \code{forecast_horizon}). T
}
\description{
Sacrifices speed somewhat for a reduction in RAM reuirements by writing
intermediary blockbuster tibbles for each timestep to disk when no longer
required.
}
\details{
See the help for \code{\link{blockbuster}} for more details.
}
\examples{

two_yrs_counterfactual <- blockbuster_sim(dplyr::filter(blockbuster_pds,
 buildingid == 127617), 2)

}
