% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{BlockbusterOutput}
\alias{BlockbusterOutput}
\alias{is.blockbuster}
\title{Defines the blockbuster class}
\usage{
BlockbusterOutput(blockbuster.output)

is.blockbuster(x)
}
\arguments{
\item{blockbuster.output}{A list with each element containing an
\code{\link{element}} object and \code{\link{block}} object.}
}
\description{
The \code{blockbuster} class is the output of the \code{\link{Blockbuster}}
function. It a list with one entry per timestep, and each list entry contains
an \code{\link{element}} object and \code{\link{block}} object providing a
snapshot of the estate condition.
}
