% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{BlockLevel}
\alias{BlockLevel}
\alias{is.block}
\title{Defines block class objects}
\usage{
BlockLevel(block.level)

is.block(x)
}
\arguments{
\item{block.level}{A data.frame containing the necessary variables.}
}
\description{
A block class object is used to hold block-level data. It identifies the
building by \code{buildingid} and contains information about the cost to
rebuild the entire block, the cost to repair all components at each grade,
and the \code{ratio} between repair and rebuild costs used to determine
which blocks are rebuilt each timestep.
}
