% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{BlockLevelList}
\alias{BlockLevelList}
\alias{is.block.list}
\title{Defines the block.list class}
\usage{
BlockLevelList(block.level.list)

is.block.list(x)
}
\arguments{
\item{block.level.list}{A list of \code{\link{block}} objects.}
}
\description{
The block.list class is used for lists of \code{\link{block}} objects.
This class is primarily used when producing plots and outputs from a
\code{\link{blockbuster}} object, that is the output from a
\code{\link{Blockbuster}} call.
}
