% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{boxplot.blockbuster_list}
\alias{boxplot.blockbuster_list}
\title{The default box-plotting of a \code{blockbuster_list} object.}
\usage{
\method{boxplot}{blockbuster_list}(x)
}
\arguments{
\item{x}{output from \code{\link{blockbuster}} function; an object of class \code{blockbuster_list}.}
}
\value{
A ggplot (boxplot) object.
}
\description{
Assigns a new method to the generic \code{boxplot} function.
}
\examples{
# boxplot(blockbuster(dplyr::filter(blockbuster_pds,
# buildingid == 127617), 5))

}

