% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{building_component_lookup}
\alias{building_component_lookup}
\title{A tibble lookup table of the \code{elementid} of each of the 139 building components in the PDS.}
\format{A dataframe of 139 rows and 4 variables.
 \describe{
 \item{elementid}{An identification number given to the type of building component.}
  \item{element}{The top level of the quantity surveyors' building component hierarchy.}
  \item{sub_element}{The mid level of the quantity surveyors' building component hierarchy.}
  \item{const_type}{The lowest level of the quantity surveyors' building component hierarchy.}
  ...
}}
\usage{
building_component_lookup
}
\description{
A dataframe containing the building component identifier \code{elementid}
 joined to its respective \code{element},
\code{sub_element} and \code{const_type}.
}
\details{
A useful lookup table when deciding upon
selecting your \code{critical_building_components} argument for \code{\link{blockbuster}}.
}
\keyword{datasets}
