% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{buildings_expected_failures}
\alias{buildings_expected_failures}
\title{Count the number of expected building failures}
\usage{
buildings_expected_failures(element, critical)
}
\arguments{
\item{element}{Data frame. Element level data. Any data frame will work as
long as it contains \code{buildingid}, \code{elementid} and \code{E} columns.}

\item{critical}{Numeric.  A vector of numbers indicating the
\code{elementid} values used to filter \code{element}.}
}
\value{
The expected number of buildings that will fail.  A failed
building is defined as one in which at least one of the critical elements has
reached grade E.  If $p_i$ is the probability that the $i$th component has
failed then the probability a building with $n$ critical components has
failed is $1-\prod_{i=1}^n\left(1-p_i\right)$.
}
\description{
Count the number of expected building failures
}
\examples{
critical_elements <- c(1700, 1701, 1706, 1707, 1708, 1752, 1756, 1769, 1773,
                       1774, 1775, 1786, 1787, 1788, 1826, 1827, 1828, 1829,
                       1830, 1883, 1885, 1887, 1889, 1892, 1893, 1900, 1903,
                       1905, 1983, 1984, 1985)
blockbuster2:::buildings_expected_failures(simulated_elements, critical_elements)
}
