% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{ConvertOutput}
\alias{ConvertOutput}
\title{Convert a block_tibble into a row per component.}
\usage{
ConvertOutput(block_tibble)
}
\arguments{
\item{block_tibble}{}
}
\value{
A tibble with one row per component, and the probability
 for each grade (A<B<C<D) as variables
}
\description{
Convert a block_tibble into a row per component.
}
