% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{ConvertPdsToWide}
\alias{ConvertPdsToWide}
\title{Convert blockbuster_pds_data sample into element level wide format}
\usage{
ConvertPdsToWide(pds, deterioration.rates = blockbuster_det_data)
}
\arguments{
\item{pds}{A slice from blockbuster_pds_data (or equivalent)}

\item{deterioration.rates}{blockbuster_det_data (or equivalent)}
}
\value{
A data.table with the appropriate rows for use by blockbuster v2
}
\description{
Deterioration rates are pulled from the default blockbuster_det_data file
and are used to populate the appropriate columns of the output.

Columns are: TODO
}
