% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_input.R
\name{create_input_element_from_excel}
\alias{create_input_element_from_excel}
\title{Loads and processes the deterioration model inputs from an excel sheet}
\usage{
create_input_element_from_excel(path = "./excel files/Excel input.xlsm")
}
\arguments{
\item{path}{character. The path to the excel sheet containing the
deterioration model inputs.}
}
\value{
A list of deterioration model inputs to pass to
 \code{\link{Blockbuster}}.  This function loads the data using the helper
 \code{\link{load_excel_inputs}} then processes the inputs (e.g. joining the
 deterioration rates and component costs)
}
\description{
Loads and processes the deterioration model inputs from an excel sheet
}
