% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CurrencyFormat}
\alias{CurrencyFormat}
\title{Find the most appropriate format for the axis of a backlog plot}
\usage{
CurrencyFormat(backlog)
}
\arguments{
\item{backlog}{A \code{\link{backlog}} class object.}
}
\value{
A list containing the magnitude by which to divide the currency,
and the suffix for the plot axis.
}
\description{
Find the most appropriate format for the axis of a backlog plot
}
