% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deteriorate.R
\name{det_eriorate}
\alias{det_eriorate}
\title{The deterioration of one blockbuster row through one time period.}
\usage{
det_eriorate()
}
\arguments{
\item{blockbuster_initial_state_row}{a blockbuster dataframe or tibble single row. 
Can only accept one row at a time due to how \code{grep} works
 in \link{/man/det_what_tm.Rd}.}
}
\value{
A one (if row is condition E) or two row tibble containing
the \code{unit_area} and condition of the \code{element sub_element constr_type} 
combination after one time period.
}
\description{
The deterioration of one blockbuster row through one time period.
}
\examples{
one_year_later <- deteriorate(blockbuster_pds[1, ])
}
\seealso{
\link{/man/blockbuster_det_data.Rd}
}

