% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_functions.R
\name{element_summarise_area}
\alias{element_summarise_area}
\title{Summarise the area at each grade for an element-level dataset}
\usage{
element_summarise_area(element_data, by = elementid)
}
\arguments{
\item{element_data}{An element-level data frame.}

\item{by}{An expression that identifies the column name used to group the
summary.  Should be either \code{buildingid} or \code{elementid}.}
}
\value{
A data frame in tidy format with grade and whatever was supplied to
the \code{by} argument as variables and the area as the value.
}
\description{
Summarise the area at each grade for an element-level dataset
}
\examples{
# Summarising area by component type
blockbuster2:::element_summarise_area(simulated_elements, elementid)
# Summarising area by block
blockbuster2:::element_summarise_area(simulated_elements, buildingid)
}
