% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{ElementLevel}
\alias{ElementLevel}
\alias{is.element}
\title{Defines element class objects}
\usage{
ElementLevel(element.level)

is.element(x)
}
\arguments{
\item{element.level}{A data.frame containing the needed variables.}
}
\description{
An element class object is used to hold element-level data.  It identifies
the building/element and contains information about the unit area, the
proportion of area at each grade, the associated unit cost of repairing each
grade, and the total cost to repair the proportion of each element at each
grade.
}
