% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{ElementLevelList}
\alias{ElementLevelList}
\alias{is.element.list}
\title{Defines the element.list class}
\usage{
ElementLevelList(element.level.list)

is.element.list(x)
}
\arguments{
\item{element.level.list}{A list of \code{\link{element}} objects.}
}
\description{
The element.list class is used for lists of \code{\link{element}} objects.
This class is primarily used when producing plots and outputs from a
\code{\link{blockbuster}} object, that is the output from a
\code{\link{Blockbuster}} call.
}
