% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{GeneratePercentageAtGrade}
\alias{GeneratePercentageAtGrade}
\title{Compute the mean proportion of elements at each grade per timestep.}
\usage{
GeneratePercentageAtGrade(input)
}
\arguments{
\item{input}{An \code{\link{element}} or \code{\link{blockbuster}} object.}
}
\value{
A table where rows are timesteps and columns are grades.
}
\description{
This function accepts both \code{\link{element.list}} and
\code{\link{blockbuster}} objects as inputs. If a \code{\link{blockbuster}}
object is provided the appropraite \code{\link{element.list}} information is
gathered using an internal call to \code{\link{PullElementData}}.
}
