% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{GenerateRepairBacklog}
\alias{GenerateRepairBacklog}
\title{Create a table containing the total repair cost of blocks at all grades and
timesteps.}
\usage{
GenerateRepairBacklog(blockbuster)
}
\arguments{
\item{block.data}{A \code{\link{block.list}}, \code{\link{block}} or
\code{\link{blockbuster}} object.}
}
\value{
A table with the grade B, C, and D repair backlogs as columns with
each row representing a timestep.
}
\description{
The input can either be a \code{\link{block.list}}, \code{\link{block}} or
\code{\link{blockbuster}} object.#'
}
