% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_functions.R
\name{InflateRebuild}
\alias{InflateRebuild}
\title{Apply inflation to the block rebuild costs in the block level data}
\usage{
InflateRebuild(block.data, inflation)
}
\arguments{
\item{block.data}{A block-level data frame.}

\item{inflation}{A number. An inflation rate of 4.5% should be entered as
1.045.}
}
\value{
A block-level data frame with the block rebuild costs inflated by the
appropriate amount.
}
\description{
Apply inflation to the block rebuild costs in the block level data
}
\examples{
# Apply 3\% inflation to rebuild costs in block-level data
blockbuster2:::InflateRebuild(simulated_blocks, 1.03)
}
