% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_functions.R
\name{InflateRepairBlock}
\alias{InflateRepairBlock}
\title{Apply inflation to the block repair costs in the block level data}
\usage{
InflateRepairBlock(block.data, inflation)
}
\arguments{
\item{block.data}{A block-level data frame.}

\item{inflation}{A number. An inflation rate of 4.5% should be entered as
1.045.}
}
\value{
A block-level data frame with the block repair costs inflated by the
appropriate amount.
}
\description{
Apply inflation to the block repair costs in the block level data
}
\examples{
# Apply 3\% inflation to block-level repair costs
blockbuster2:::InflateRepairBlock(simulated_blocks, 1.03)
}
