% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_functions.R
\name{InflateRepairElement}
\alias{InflateRepairElement}
\title{Apply inflation to the element repair costs in the element level data}
\usage{
InflateRepairElement(element.data, inflation)
}
\arguments{
\item{element.data}{An element-level data frame.}

\item{inflation}{A number. An inflation rate of 4.5% should be entered as
1.045.}
}
\value{
An element-level data frame with the repair costs and totals inflated
by the appropriate amount.
}
\description{
Apply inflation to the element repair costs in the element level data
}
\examples{
# Apply 3\% inflation to the element-level repair costs and totals.
blockbuster2:::InflateRepairElement(simulated_elements, 1.03)
}
