% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_input.R
\name{load_excel_inputs}
\alias{load_excel_inputs}
\title{Loads deterioration model inputs from the input excel sheet}
\usage{
load_excel_inputs(path)
}
\arguments{
\item{path}{character. The path to the input excel sheet containing the
deterioration model inputs}
}
\value{
A list of inputs to pass to the
 \code{\link{create_input_element_from_excel}} function
}
\description{
Loads deterioration model inputs from the input excel sheet
}
