% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{LoadBlockbusterOutput}
\alias{LoadBlockbusterOutput}
\title{Load data from the blockbuster output files.}
\usage{
LoadBlockbusterOutput(forecast.horizon, filelabel = "blockbuster_output",
  path = "./output/")
}
\arguments{
\item{forecast.horizon}{The number of timesteps to collate.}

\item{filelabel}{Character. The file label on the output files. This
corresponds to the \code{filelabel} argument passed to \code{Blockbuster}.
Default is "blockbuster_output".}

\item{path}{Character. The path to the folder containing the saved files.
This corresponds to the \code{path} argument passed to \code{Blockbuster}.
Default is "./output/".}
}
\value{
A list of lists.  There is one list entry for each timestep, so
the output will be of length \code{forecast.horizon + 1} as the first will
contain the initial state.  Each list entry contains a block-level and
element-level data frame named \code{block} and \code{element} respectively.
}
\description{
This function loads the blockbuster outputs into memory from the saved files
produced during a call to \code{\link{Blockbuster}}.
}
\details{
The files are expected to end in "_element_i" and "block_i", with i
being the timestep plus one. If the data input to \code{\link{Blockbuster}}
is large then R may run out of memory when creating the final output.  For
this reason the state of the estate can be saved to disc and the final output
reconstructed at the end of the process, or outside of the
\code{\link{Blockbuster}} call.
}
\examples{
\dontrun{
# Load the first ten years of output from the default folder
LoadBlockbusterOutput(10)
}
}
