% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_input.R
\name{log_line}
\alias{log_line}
\title{Writes to a file and include a timestamp}
\usage{
log_line(x, log_file = log_file)
}
\arguments{
\item{x}{character.  The text to write}

\item{log_file}{A file connection}
}
\value{
No return.
}
\description{
A helper function intended for generating a text log.  It writes a text line
to a file, and includes a time-stamp#'
}
