% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blockbuster.R
\name{MultiBlockbuster}
\alias{MultiBlockbuster}
\title{Run Blockbuster simulations on a variety of spending strategies.}
\usage{
MultiBlockbuster(strat, scenarios, element.data, block.data, forecast.horizon,
  inflation, block.rebuild.cost)
}
\arguments{
\item{strat}{Character.  A single character string corresponding with one
of the entries in the strategy column of \code{scenarios}.}

\item{scenarios}{Data frame. Strategies are named in the last column and each
strategy is associated with two rows, the first containing the rebuild budget,
the second containing the repair budget.}

\item{element.data}{An \code{\link{element}} object containing the initial
state.}

\item{block.data}{A \code{\link{block}} object containing the initial state.}

\item{forecast.horizon}{A number indicating the number of timesteps to
simulate.}

\item{inflation}{A vector of numbers indicating the inflation rate to apply
to repair and rebuild costs each timestep.  If a vector of length one is
supplied it will be used as the inflation rate for all timesteps.}

\item{block.rebuild.cost}{The unit cost (per m^2^) of rebuilding blocks.}
}
\value{
NULL.  This function does not return an object, but instead saves all
outputs (and interim states) to files in the "./output/" folder.
}
\description{
Given a data.frame with of spending scenarios, run the scenario with the
\code{strat} argument in the strategy column.  This function is useful for
passing to lapply with a list of strategy names as the first argument.
}
