% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{PercentageAtGrade}
\alias{PercentageAtGrade}
\title{Compute the mean proportion of building components at each grade.}
\usage{
PercentageAtGrade(element.data)
}
\arguments{
\item{element.data}{An \code{\link{element}} object.}
}
\value{
A data table containing the mean proportions at each grade.
}
\description{
Compute the mean proportion of building components at each grade.
}
