% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.blockbuster_list}
\alias{plot.blockbuster_list}
\title{The default plotting of a \code{blockbuster_list} object.}
\usage{
plot.blockbuster_list(x, ...)
}
\arguments{
\item{blockbuster_list}{output from \code{\link{blockbuster}} function; an object of class \code{blockbuster_list}.}
}
\value{
A ggplot object (lineplot).
}
\description{
Assigns a new method to the generic \code{plot} function.
}
\details{
Produces a line plot of total cost of repairs by grade through time.
}
\examples{
plot(blockbuster(dplyr::filter(blockbuster_pds,
 buildingid == 127617), 5))

}

