% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{plot.repair.backlog}
\alias{plot.repair.backlog}
\title{Plot method for repair.backlog objects}
\usage{
\method{plot}{repair.backlog}(backlog, total = FALSE, grades = c("C", "D"))
}
\arguments{
\item{backlog}{A \code{\link{repair.backlog}} object.}

\item{total}{Logical. (optional) Default is \code{FALSE}. If \code{TRUE} then the
combined C,D backlog is plotted as a timeseries}
}
\description{
If the input only covers one timestep then a barchart is plotted. Otherwise a
time series of the backlog at each grade is plotted.
}
