% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{pull_Element_Data}
\alias{pull_Element_Data}
\alias{pull_Block_Data}
\alias{load_Element_Data}
\alias{load_Block_Data}
\title{Collate element- and block-level outputs from Blockbuster outputs}
\usage{
pull_Element_Data(blockbuster_output)

pull_Block_Data(blockbuster_output)

load_Element_Data(forecast.horizon, path = "./output/",
  filelabel = "blockbuster_output")

load_Block_Data(forecast.horizon, path = "./output/",
  filelabel = "blockbuster_output")
}
\arguments{
\item{blockbuster_output}{The output from a \code{\link{Blockbuster}} call.}

\item{forecast.horizon}{integer.  This specifies how many years of the saved
output will be loaded.}

\item{path}{character. This should be the same as the argument passed to the \code{\link{Blockbuster}} function.}

\item{filelabel}{character. This should be the same as the argument passed to the \code{\link{Blockbuster}} function.}
}
\value{
A data.frame containing the appropriate data from the output, with a year column
}
\description{
A set of convenience functions for managing the raw output of a \code{\link{Blockbuster}} function call.
}
\details{
The functions \code{pull_Block_data} and \code{pull_Element_data} combine the appropriate data into a single dataframe with a column for year.

For large datasets (for example, the PDS dataset), R can hit memory limits when the \code{\link{Blockbuster}} function tries to construct the single output object from the saved interim files. \code{load_Block_Data} and \code{load_Element_Data} loads the relevant outputs from the saved files.
}
\section{Functions}{
\itemize{
\item \code{pull_Element_Data}: Pulls out the element level data from the output of a blockbuster function call

\item \code{pull_Block_Data}: Pulls out the block data from the output from a blockbuster function call

\item \code{load_Element_Data}: Loads the element-level output from a blockbuster call

\item \code{load_Block_Data}: Loads the element-level output from a blockbuster call
}}

\examples{
\dontrun{
output <- Blockbuster(simulated_elements, forecast_horizon = 10)
pull_Block_Data(output, 10)
pull_Element_Data(output, 10)
}
# loading outputs from saved files
\dontrun{
Blockbuster(simulated elements, path = "output", filelabel = "example")
load_Element_data(1, path = "output", filelabel = "example")
load_Block_data(1, path = "output", filelabel = "example")
}
}
