% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{PullBlockData}
\alias{PullBlockData}
\title{Creates a block.list object from a blockbuster object}
\usage{
PullBlockData(blockbuster)
}
\arguments{
\item{blockbuster}{A \code{\link{blockbuster}} object.}
}
\value{
A list of block-level information as a \code{\link{block.list}}
object
}
\description{
Creates a list of class \code{\link{block.list}} containing all the block-
level information from the output from a \code{\link{Blockbuster}} model.
}
