% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{PullElementData}
\alias{PullElementData}
\title{Creates an element.list object from a blockbuster object}
\usage{
PullElementData(blockbuster)
}
\arguments{
\item{blockbuster}{A \code{\link{blockbuster}} object.}
}
\value{
A list of element-level information as an \code{\link{element.list}}
object
}
\description{
Creates a list of class \code{\link{element.list}} containing all the
element-level information from the output from a \code{\link{Blockbuster}}
model.
}
