% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_input.R
\name{render_blockbuster}
\alias{render_blockbuster}
\title{Render output word document from output excel}
\usage{
render_blockbuster(input, output, file)
}
\value{
Generates a word document
}
\description{
Render output word document from output excel
}
