% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes_and_methods.R
\name{RepairBacklogClass}
\alias{RepairBacklogClass}
\title{Defines the repair.backlog class}
\usage{
RepairBacklogClass(backlog)
}
\arguments{
\item{backlog}{A data frame containing columns titled \code{timestep},
\code{grade} and \code{backlog}. There can be other columns.}
}
\description{
The repair.backlog class is a long data format holding repair backlogs
suitable for passing to \code{ggplot2}.
}
