% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_functions.R
\name{RepairBlock}
\alias{RepairBlock}
\title{Repair components in element data when there is not enough money to repair
them all.}
\usage{
RepairBlock(element.data, repair.money)
}
\arguments{
\item{element.data}{An \code{\link{element}} class object.}

\item{repair.money}{A number.}
}
\value{
An \code{\link{element}} class object with the grade proportions
amended accordingly.
}
\description{
This function is called by \code{\link{Repair}} when it needs to make
decisions about what to spend money on repairing.  Despite the name, it can
be applied to element-level data that contains components from more than one
block. A greedy algorithm is applied to spend the available repair money.
Components are repaired in order of descending cost by grade, starting with
grade D, then C, then B.
}
\details{
This function does NOT update the repair costs, only the grade proportions.
Therefore \code{\link{UpdateElementRepairs}} and
\code{\link{UpdateBlockRepairs}} should be called on the element- and block-
level objects afterwards to ensure the repair costs and totals are correct.
}
