% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_functions.R
\name{repairComponent}
\alias{repairComponent}
\title{Repairs the component in the given rows at the given grade}
\usage{
repairComponent(element.data, rows, grade)
}
\arguments{
\item{element.data}{data.frame. The component level data.}

\item{grade}{character. One of \code{"B"}, \code{"C"}, \code{"D"}, \code{"E"}.}

\item{row}{numeric.  A vector of row indices to update}
}
\value{
The updated element.data
}
\description{
Repairs the component in the given rows at the given grade
}
