% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_functions.R
\name{RepairOld}
\alias{RepairOld}
\title{Repair elements in a block}
\usage{
RepairOld(element.data, block.data, repair.money)
}
\arguments{
\item{element.data}{An \code{\link{element}} class object.}

\item{block.data}{A \code{\link{block}} class object.}

\item{repair.money}{A number}
}
\value{
An \code{\link{element}} class object with relevant components
repaired, i.e. set to grade A.
}
\description{
Given the current state of the buildings, this function outputs the element-
level status after repairing components within budget.
}
\details{
The available funding for repairs, input as \code{repair.money}, is split
evenly between blocks.  Within each block the components are repaired in
descending order of cost. Note that this method can result in a large amount
of wasted fund allocation.  For example, a block with only N or A grade
components will be allocated a proportion of the funds, but needs none.

This function does NOT update the repair costs, only the grade proportions.
Therefore \code{\link{UpdateElementRepairs}} and
\code{\link{UpdateBlockRepairs}} should be called on the element- and block-
level objects afterwards to ensure the repair costs and totals are correct.
}
