% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_blocks}
\alias{simulated_blocks}
\title{Small toy block-level dataset}
\format{A dataframe with 50 rows and 7 variables:
\describe{
  \item{buildingid}{A key uniquely identifiying the block}
  \item{block.rebuild.cost}{The cost of rebuilding the block}
  \item{B.block.repair.cost}{The expected cost of repairing all components in
    the block at grade B}
  \item{C.block.repair.cost}{The expected cost of repairing all components in
    the block at grade C}
  \item{D.block.repair.cost}{The expected cost of repairing all components in
    the block at grade D}
  \item{E.block.repair.cost}{The expected cost of repairing all components in
    the block at grade E}
  \item{ratio}{The efficiency of rebuilding rather than repairing}
}}
\usage{
simulated_blocks
}
\description{
A simulated block-level dataset compatible with \code{\link{Blockbuster}}. It
contains simulated data for 50 buildings and is a summary of the data in
\code{\link{simulated_elements}}
}
\keyword{datasets}
