% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{single_backlog}
\alias{single_backlog}
\title{Pull the backlog column for a given grade from block data}
\usage{
single_backlog(block, grade)
}
\arguments{
\item{block}{A data frame containing block-level data.}

\item{grade}{Character.  Either \code{"B"}, \code{"C"}, \code{"D"}, or
\code{"E"}. Anything else will return an error.}
}
\value{
A vector of the repair costs for the appropriate grade.
}
\description{
Pull the backlog column for a given grade from block data
}
