% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.blockbuster_list}
\alias{summary.blockbuster_list}
\title{The default summary of a \code{blockbuster_list} object.}
\usage{
summary.blockbuster_list(x)
}
\arguments{
\item{x}{output from \code{\link{blockbuster}} function; an object of class \code{blockbuster_list}.}
}
\value{
A list containing some descriptive text and
}
\description{
Assigns a new method to the generic \code{summary} function.
}
\details{
Provides an overview of total costs by year and the proportion of the unit_area in
the different grades for the final year of the simulation.
}
\examples{
# summary(blockbuster(dplyr::filter(blockbuster_pds,
 # buildingid == 127617), 5))

}
