% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.R
\name{what_needs_repair_within_block}
\alias{what_needs_repair_within_block}
\title{A helper function that updates the \code{repair_status} of building components in a block.}
\usage{
what_needs_repair_within_block(block_tibble, per_block_spend)
}
\arguments{
\item{block_tibble}{a block dataframe or tibble with a \code{costs}variable.}

\item{per_block_spend}{a scalar calculated from \code{repair_monies} / number of blocks.}
}
\value{
A \code{block_tibble} that has had building components repaired or not indicated
by the new variable \code{repair_status}.
}
\description{
Outputs a block_tibble after spending \code{per_block_spend} on one block. Used in the 
\code{\link{repair}} function. A new binary variable is created of the \code{repair_status} 
of each building component within the block. The building components are sorted by 
descending grade and then descending cost in the \code{\link{repair}} function prior to 
being passed to this helper function.
}
\examples{

example <- what_needs_repair_within_block(tibble::tibble(
cost = seq(from = 0, to = 10000, by = 500)),
 1e4)
}

