% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backlog_functions.R
\name{backlog}
\alias{backlog}
\title{Sums the backlog in block data}
\usage{
backlog(block, grade = c("B", "C", "D", "E"))
}
\arguments{
\item{block}{A \code{\link{block}} class object.}

\item{grade}{Character vector. Elements in the vector must be either
\code{"B"}, \code{"C"}, \code{"D"}, or \code{"E"}. Anything else will return
 an error. Set as \code{c("B", "C", "D", "E")} by default.}
}
\value{
Numeric. The sum of the repair costs for all grades specified in
 \code{grade}.
}
\description{
Sums the backlog in block data
}
\examples{
\dontrun{
# Total grade B backlog
backlog(block_data, "B")
# Total grade B and C backlog
backlog(block_data, c("B", "C"))
# All backlog
backlog(block_data)
}
}
