% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deteriorate.R
\name{blockbuster}
\alias{blockbuster}
\title{The deterioration of more than one blockbuster rows through one time period.}
\usage{
blockbuster(blockbuster_tibble, forecast_horizon)
}
\arguments{
\item{blockbuster_tibble}{a blockbuster dataframe or tibble.}

\item{forecast_horizon}{an integer for the number of timesteps to model deterioration over.}
}
\value{
A list of n tibbles (where n is based on the \code{forecast_horizon}),
with each tibble containing
the \code{unit_area} and condition of the \code{element sub_element constr_type} 
combination at a given timestep while also duplicating all
 other variables and values from the input tibble.
After each timestep is simulated the rows are aggregated by \code{elementid} and \code{grade}.
}
\description{
The deterioration of more than one blockbuster rows through one time period.
}
\examples{
two_year_later <- blockbuster(dplyr::filter(blockbuster_pds, buildingid == 127617), 2)

}

