% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockbuster_det_data}
\alias{blockbuster_det_data}
\title{Deterioration rates of all blockbuster element-sub-element-construction types.}
\format{A tibble with 147 rows and 9 variables:
\describe{
  \item{element}{Building taxonomic classification}
  \item{sub_element}{}
  \item{const_type}{Construction type}
  \item{transition rates}{From a to b, b to c, c to d and d to e.}
  \item{concated_det}{A pasting together of the element sub-element and const_type; 
  keeping alphanumeric and spaces only.}
  ...
}}
\usage{
blockbuster_det_data
}
\description{
A dataset containing the Deterioration rates of all element by sub-element
by construction type combinations. These rates were provided by the
consultant EC Harris and are necessary for predicting the
deterioration of each element by sub-element
by construction type combination through time as in the
 \code{det_eriorate} function in this package.
}
\keyword{datasets}

