% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blockbuster_pds}
\alias{blockbuster_pds}
\title{A random ten percent sample of the condition of the School Estate.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 254878 rows and 25 columns.}
\source{
\url{https://www.gov.uk/government/organisations/education-funding-agency}
}
\usage{
blockbuster_pds
}
\description{
A dataset containing the condition grade and other attributes of almost 255,000
building components of the English School Estate. The unit_area was estimated using
the \code{areafy2} function in this package. School identifying features have been removed.
}
\keyword{datasets}

