% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcoster.R
\name{blockcoster_lookup}
\alias{blockcoster_lookup}
\title{A lookup table for repair cost rate given a building component and its grade.}
\usage{
blockcoster_lookup(the_elementid, the_grade,
  costs_lookup = blockbuster_pds_repair_costs)
}
\arguments{
\item{the_elementid}{the building component unique identifier found in the PDS.}

\item{the_grade}{the grade of a blockbuster tibble single row (a factor with 6 levels).}

\item{costs_lookup}{the relevant costs look up table, default is derived from PDS 2016 costs}
}
\value{
a numeric constant for the repair cost of building component by grade, in pounds per unit_area.
}
\description{
An internal function in \code{\link{blockbuster}} to select the correct repair cost constant for a building component and its condition grade. 
It works by matching on the concatenated \code{elementid} and \code{grade}.
}
\details{
The default is to use the repair cost estimates for the PDS plus
an E grade cost estimator based on D grade cost plus 5% (removal costs etc.). This
is likely to be a source of error as the replacement of a decommisioned component
depends on the component as well as other complexities. This should be addressed in later
improvements of the \code{\link{blockbuster}}.
}
\examples{
x <- blockcoster_lookup(
blockbuster_pds[1:10, ]$elementid,
 blockbuster_pds[1:10, ]$grade)
}
\seealso{
\code{\link{blockbuster_pds_repair_costs}}
}

