% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockcoster.R
\name{blockcoster_lookup}
\alias{blockcoster_lookup}
\title{An internal function in \code{\link{blockbuster}} to select the correct repair cost constant for a building component and its condition grade. 
It works by matching strings from the concatenated \code{element}, \code{sub_element}, \code{const_type} and \code{grade}.}
\usage{
blockcoster_lookup(element, sub_element, const_type, grade,
  costs_lookup = blockbuster_pds_repair_costs)
}
\arguments{
\item{element}{the element of a blockbuster tibble single row.}

\item{sub_element}{the sub_element of a blockbuster tibble single row}

\item{const_type}{the const_type of a blockbuster tibble single row}

\item{grade}{the grade of a blockbuster tibble single row}

\item{costs_lookup}{the relevant costs look up table, default is derived from PDS 2016 costs}
}
\value{
a numeric constant for the repair cost of building component by grade, in pounds per unit_area.
}
\description{
An internal function in \code{\link{blockbuster}} to select the correct repair cost constant for a building component and its condition grade. 
It works by matching strings from the concatenated \code{element}, \code{sub_element}, \code{const_type} and \code{grade}.
}
\examples{
cost_to_repair_one_unit_area_back_to_grade_a <- blockcoster_lookup(
blockbuster_pds[3, ]$element, blockbuster_pds[3, ]$sub_element, 
blockbuster_pds[3, ]$const_type, blockbuster_pds[3, ]$grade)
}
\seealso{
\code{\link{blockbuster_pds_repair_costs}}
}
