% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion_functions.R
\name{ConvertPdsToBlock}
\alias{ConvertPdsToBlock}
\title{Convert blockbuster_pds_data sample into block level wide format}
\usage{
ConvertPdsToBlock(pds, block.rebuild.cost)
}
\arguments{
\item{pds}{blockbuster_det_data (or equivalent)}

\item{block.rebuild.cost}{A number}

\item{pds}{blockbuster_det_data (or equivalent)}

\item{block.rebuild.cost}{A number}
}
\value{
A block level wide format table for use with blockbuster v2
Convert blockbuster_pds_data sample into block level wide format

A block level wide format table for use with blockbuster v2
}
\description{
Convert blockbuster_pds_data sample into block level wide format
}
\details{
block.rebuild.cost is used to compute the estimated cost of rebuilding
each building as block.rebuild.cost x gifa (Gross Internal Floor Area).

block.rebuild.cost is used to compute the estimated cost of rebuilding
each building as block.rebuild.cost x gifa (Gross Internal Floor Area).
}
