% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deteriorate.R
\name{det_eriorate}
\alias{det_eriorate}
\title{The deterioration of one blockbuster row through one time period.}
\usage{
det_eriorate(blockbuster_initial_state_row)
}
\arguments{
\item{blockbuster_initial_state_row}{a blockbuster dataframe or tibble single row. 
Can only accept one row at a time due to how \code{grep} works
 in \code{\link{det_what_tm}}.}
}
\value{
A one (if row is condition E) or two row tibble containing
the \code{unit_area} and condition of the \code{element sub_element constr_type} 
combination after one time period. Duplicating all other variables and values.
The timestep needs to increase by one.
}
\description{
The deterioration of one blockbuster row through one time period.
}
\examples{
one_year_later <- det_eriorate(blockbuster_pds[1, ])
}
\seealso{
\code{\link{blockbuster_det_data}}
}

