% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deteriorate.R
\name{det_what_tm}
\alias{det_what_tm}
\title{Internal function that selects the correct markovchain object for deterioration of a row.}
\usage{
det_what_tm(blockbuster_initial_state_row)
}
\arguments{
\item{blockbuster_initial_state_row}{a blockbuster dataframe or tibble single row.}
}
\value{
a markovchain object containing the appropriate transition matrix
for the input blockbuster tibble row. Can only accept one row at a time due to how grep works.
}
\description{
New approach that should improve speed 
because testing equality (of \code{elementid}) is simpler than testing
 inclusion in a set using %in%.
}
\examples{
old_strings_method <- det_what_tm(
dplyr::mutate_(blockbuster_pds[1, ], 
concated = ~gsub(pattern = "[^[:alnum:] ]",
replacement = "",
paste(element, sub_element, const_type,
      sep = ""))))  #  old string method lookup required concated cached
      
new_lookup_is_faster <- det_what_tm(blockbuster_pds[96, ])
      
}
\seealso{
\code{\link{blockbuster_det_data}}
}
