% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deteriorate_function.R
\name{Deteriorate}
\alias{Deteriorate}
\title{Deteriorate all the elements in an element-level data frame.}
\usage{
Deteriorate(element.data)
}
\arguments{
\item{element.data}{An element-level data frame.}
}
\value{
An element-level data frame.
}
\description{
Takes an element-level data frame and amends the grade proportions for each
row according to the deterioration rate in each row. This function is used by
the \code{\link{Blockbuster}} function to compute the deterioration at each
timestep.
}
\details{
The deterioration rates are based on lifetime estimates from EC Harris and
other expert opinion.
}
\examples{
blockbuster2:::Deteriorate(simulated_elements)
}
