% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_functions.R
\name{InflateRepairBlock}
\alias{InflateRepairBlock}
\title{Apply inflation to the block repair costs in the block level data}
\usage{
InflateRepairBlock(block.data, inflation)
}
\arguments{
\item{block.data}{A \code{\link{block}} class object.}

\item{inflation}{A number.}
}
\value{
A \code{\link{block}} class object with the block repair costs
inflated by the appropriate amount.
}
\description{
An inflation rate of 4.5% should be entered as 1.045.
}
