% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_functions.R
\name{InflateRepairElement}
\alias{InflateRepairElement}
\title{Apply inflation to the element repair costs in the element level data}
\usage{
InflateRepairElement(element.data, inflation)
}
\arguments{
\item{element.data}{An \code{\link{element}} class object.}

\item{inflation}{A number.}
}
\value{
An \code{\link{element}} class object with the repair costs and
totals inflated by the appropriate amount.
}
\description{
An inflation rate of 4.5% should be entered as 1.045.
}
