% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{rebuild}
\alias{rebuild}
\title{An internal function that determines rebuilding intervention effect on a blockbuster_tibble.}
\usage{
rebuild(blockbuster_tibble, rebuild_monies)
}
\arguments{
\item{blockbuster_tibble}{a blockbuster dataframe or tibble.}

\item{rebuild_monies}{a vector of length one.}
}
\value{
A \code{blockbuster_tibble} that has had blocks rebuilt (or not if \code{rebuild_monies} are 0).
}
\description{
A given amount of money (\code{rebuild_monies}) is invested in rebuilding
 the \code{blockbuster_tibble}, passed as an arguement to the function. 
 This is dependent on the \code{block_rebuild_cost} and the \code{rebuild_monies}.
}
\details{
Outputs a blockbuster_tibble after spending \code{rebuild_monies} on a finite
number of blocks. Rebuilding a block converts all it's building components
back to grade N, making it as new. The rebuilding decision making is based on
a summarised sorted list of blocks, sorted by the ratio of the total cost of repairs
divided by the rebuild cost of the block (this is calculated in \code{\link{blockbuster}}).
 As this ratio approaches one it suggests the
cost of repairs are equivalent to building a new block! Using the available funding
or \code{rebuild_monies} we systematically work through and rebuild each block based
on our sorted list until all the money is used up. If a block is to expensive to rebuild,
we move on and attempt to rebuild the next one. If our \code{rebuild_monies} is less than
the cheapest block to repair we stop trying to rebuild. Thus money can be wasted. Upon 
rebuilding the \code{unit_area} must be re-estimated for each building component thus the 
\code{\link{areafy2}} function is called internally. If \code{rebuild_monies} are zero
(the default) then the rebuilding is skipped and the \code{blockbuster_tibble} argument
is the output.
}
\examples{

jackpot <- blockbuster(dplyr::filter(blockbuster_pds,
  buildingid == 4382 | buildingid == 4472 | buildingid == 4487), 
  forecast_horizon = 2, rebuild_monies = 5e6)

}
