% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_functions.R
\name{Rebuild}
\alias{Rebuild}
\title{Rebuild blocks by adjusting the element object}
\usage{
Rebuild(element.data, block.data, rebuild.money)
}
\arguments{
\item{rebuild.money}{}
}
\value{
An \code{\link{element}} object with updated grade proportions for
rebuilt blocks.  Note that repair costs are not updated, nor the
\code{\link{block}} object holding block-level information so
\code{\link{UpdateElementRepairs}} and \code{\link{UpdateBlockRepairs}}
should be run afterwards.
}
\description{
This function simulates rebuilding by changing all component blocks to grade
N in rebuild blocks.  The output is the element-level object of class
\code{\link{element}} and the repair costs are not updated in this function.
It is therefore important to run \code{\link{UpdateElementRepairs}} and
\code{\link{UpdateBlockRepairs}} on the \code{\link{element}} and
\code{\link{block}} objects respectively after calling this function, and in
that order too.
}
\details{
The decision about which blocks to rebuild is made based on a measure of
cost-effectiveness, the ratio between the cost to repair all components in a
block and the cost to rebuild the entire block.  The rebuild cost is
estimated using the unit rebuild cost times the Gross Internal Floor Area
(GIFA) of the block.

After eliminating all zero rebuild cost block from consideration (to save
computational time), the blocks are ordered in descending order of
cost effectiveness.  If there is enough money then the first block is
repaired, then the second, and so on, skipping blocks if there are not
enough funds.  This continues until either there are no blocks left to
rebuild or not enough money to rebuild any remaining blocks.
}
