% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.R
\name{repair}
\alias{repair}
\title{An internal function that determines repairs intervention effect on a blockbuster_tibble.}
\usage{
repair(blockbuster_tibble, repair_monies)
}
\arguments{
\item{blockbuster_tibble}{a blockbuster dataframe or tibble.}

\item{repair_monies}{a vector of length one.}
}
\value{
A \code{blockbuster_tibble} that has had building components repaired
 (or not if \code{repair_monies} are 0).
}
\description{
A given amount of money (\code{repair_monies}) is invested in repairing
 the \code{blockbuster_tibble}, passed as an arguement to the function. 
 The \code{repair_monies} is distributed evenly to each \code{buildingid} 
 included in the \code{blockbuster_tibble}. Grade D, C and B condition 
 building components are repaired to A in that order. Partial repairs does not happen
 and any overspend is discarded (spent by the school on other things).
}
\details{
Outputs a blockbuster_tibble after spending \code{repair_monies} on a finite
number of blocks. Repairing a block attempts to convert some 
(depending on \code{repair_monies} available) of it's building components
back to grade A, making it Excellent condition. If the building components have
zero cost they are assumed to be grade N or A (this is not always true) and are not
repaired (the cost model is the issue not the assumption). The input is seperated into
a list of dataframes by block and then \code{repair_status} binary variable added using 
the internal helper function \code{\link{what_needs_repair_within_block}}. The repairs are 
then carried out an additional helper function called ...
}
\examples{

quick_fix <- blockbuster(dplyr::filter(blockbuster_pds,
  buildingid == 4382 | buildingid == 4472 | buildingid == 4487), 
  forecast_horizon = 2, rebuild_monies = 0, repair_monies = 2.5e6)

}
