% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_functions.R
\name{repairComponent}
\alias{repairComponent}
\title{Repairs the component in the given rows at the given grade}
\usage{
repairComponent(element.data, rows, grade)
}
\arguments{
\item{grade}{}

\item{row}{numeric.  A vector of row indices to update}
}
\value{
The updated element.data
}
\description{
Repairs the component in the given rows at the given grade
}
