% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_functions.R
\name{repairGrade}
\alias{repairGrade}
\title{Repair components of a particular grade}
\usage{
repairGrade(element.data, grade)
}
\arguments{
\item{element.data}{An \code{\link{element}} object, with a repair_money attribute indicating the money allowed for repairs}

\item{grade}{Either \code{"B"}, \code{"C"}, \code{"D"} or \code{"E"}.  The grade to
repair}

\item{budget}{Numeric. The amount of funds available for repair.}
}
\value{
The \code{\link{element}} object with repairs
completed, and a numeric amount stating the remaining funds as an attribute.
}
\description{
This function is used by \code{\link{Repair}} to repair all components at a
particular grade.
}
\details{
Components are sorted by their proportion at the given grade.  They are then
repaired in descending order until the money given by \code{budget} runs out.
}
