% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_functions.R
\name{repairGrade}
\alias{repairGrade}
\title{Repair components of a particular grade}
\usage{
repairGrade(element.data, repair.money, grade)
}
\arguments{
\item{element.data}{An \code{\link{element}} object.}

\item{grade}{Either \code{"B"}, \code{"C"} or \code{"D"}.  The grade to
repair}

\item{budget}{Numeric. The amount of funds available for repair.}
}
\value{
A list containing the \code{\link{element}} object with repairs
completed, and a numeric amount stating the remaining funds.
}
\description{
This function is used by \code{\link{Repair}} to repair all components at a
particular grade.
}
\details{
Components are sorted by their proportion at the given grade.  They are then
repaired in descending order until the money given by \code{budget} runs out.
}
